requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# NLP and topic modeling
gensim>=4.0.0
spacy>=3.0.0

# Data processing
scikit-learn>=1.0.0

# Configuration
PyYAML>=6.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Optional: Progress bars
tqdm>=4.60.0
